feat: moving macOS settings - [2/20]
- add a new module `settings` in the src/platform/macos/mod.rs file
- implement parsing and conversion functions for `optionasmeta` enum in the new src/platform/macos/settings.rs file
- refactor the src/window/settings.rs file to remove duplicate code for `optionasmeta` enum and related functions
